requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
databricks-sdk>=0.30.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
python-multipart>=0.0.9
opencv-python-headless>=4.9.0
Pillow>=10.0.0
